Moved the Gateways all into the DataRetrieval folder Wrote javadocs for both & finalized the functionality
